Added gun and Sprinting

+++ FOGGS_OpenGL_FreeGLUT/FOGGS_OpenGL_FreeGLUT/BMPLoader.cpp
@@ -22,11 +22,10 @@ namespace BMPLoader {
 			unsigned char info[54] = { 0 };
 			inFile.read((char*)&info[0], 54);
 
-			//Ask about "*(unsigned int*)&", without the *'s and & symbol it returns 0.
 			bmp->width = *(unsigned int*)&info[18];
 			bmp->height = *(unsigned int*)&info[22];
 
-			unsigned long long size = ((((bmp->width * bmp->height) + 31) & ~31) / 8) * bmp->height;
+			unsigned long size = ((((bmp->width * bmp->height) + 31) & ~31) / 8) * bmp->height;
 			bmp->pixels = new unsigned char[size];
 			inFile.read((char*)bmp->pixels, size);
 			inFile.close();
